Hopefully done with colum solver up to last 2
The column solver is good enough for now up to the last 2 cells. The
column solver assumes that, when starting a new column, all cells it
needs to move are in the same column or a column to the left. Because
the last two cells are left unsolved, the condition is not guaranteed to
be true. Hence, to properly test the column solving, I need to solve the
last two cells