Add support for special cases of the last 2 cells in column
When the last 2 cells in a column are in the top right 2x2 of the
unsolved board but are not in place, trying to put them in place by the
algorithm I would use would not work. Therefore, they are a special case
that must be handled.
Wrote effic.py to generate a shortest list of moves for each case. It
it is not a dependency for execution/compilation.