Greedy ai prototype 2
The algorithm to complete the last two cells is sort of complete. It
makes a uses a assumption that 0 always starts out either above or to
the left of the 3rd to last cell in a column which I'll need to fix.
It would also probably segfault eventually even if that assumption were
true, I haven't tested it much.